+++ packages/cli/generation/ir-generator/src/generateIntermediateRepresentation.ts
@@ -188,7 +188,8 @@ export function generateIntermediateRepresentation({
                       docs: typeof variable !== "string" ? variable.docs : undefined,
                       id: key,
                       name: rootApiFileContext.casingsGenerator.generateName(key),
-                      type: rootApiFileContext.parseTypeReference(variable)
+                      type: rootApiFileContext.parseTypeReference(variable),
+                      default: typeof variable !== "string" ? (variable.default ?? undefined) : undefined
                   }))
                 : [],
         serviceTypeReferenceInfo: {